backend/requirements.txt
asgiref==3.8.1
astroid==3.3.4
autopep8==2.3.1
black==24.8.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
dill==0.3.8
Django==5.1.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
idna==3.10
isort==5.13.2
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.3.6
psycopg2==2.9.9
pycodestyle==2.12.1
pylint==3.3.1
requests==2.32.3
sqlparse==0.5.1
tomlkit==0.13.2
urllib3==2.2.3